+++ src/avs/AVSDetails.jsx
@@ -10,7 +10,7 @@ import { useLocation } from 'react-router-dom';
 import { useMutativeReducer } from 'use-mutative';
 import { useParams } from 'react-router-dom';
 import { useServices } from '../@services/ServiceContext';
-import { useTailwindBreakpoint } from '../shared/useTailwindBreakpoint';
+import { useTailwindBreakpoint } from '../shared/hooks/useTailwindBreakpoint';
 
 export default function AVSDetails() {
   const { address, tab } = useParams();
