+++ src/avs/AVSDetailsTVLTab.jsx
@@ -26,7 +26,7 @@ import TVLTabTreemap from './charts/TVLTabTreemap';
 import { useMutativeReducer } from 'use-mutative';
 import { useParams } from 'react-router-dom';
 import { useServices } from '../@services/ServiceContext';
-import { useTailwindBreakpoint } from '../shared/useTailwindBreakpoint';
+import { useTailwindBreakpoint } from '../shared/hooks/useTailwindBreakpoint';
 
 export default function AVSDetailsTVLTab({
   totalTokens,
