+++ src/operators/OperatorDetails.jsx
@@ -18,7 +18,7 @@ import { useEffect } from 'react';
 import { useMutativeReducer } from 'use-mutative';
 import { useParams } from 'react-router-dom';
 import { useServices } from '../@services/ServiceContext';
-import { useTailwindBreakpoint } from '../shared/useTailwindBreakpoint';
+import { useTailwindBreakpoint } from '../shared/hooks/useTailwindBreakpoint';
 
 export default function OperatorDetails() {
   const { address } = useParams();
